+++ lib/modules/anime/widgets/desktop.dart
@@ -500,75 +500,6 @@ class _DesktopControllerWidgetState extends State<DesktopControllerWidget> {
   }
 }
 
-// BUTTON: PLAY/PAUSE
-
-/// A material design play/pause button.
-class CustomeMaterialDesktopPlayOrPauseButton extends StatefulWidget {
-  final VideoController controller;
-
-  const CustomeMaterialDesktopPlayOrPauseButton({
-    super.key,
-    required this.controller,
-  });
-
-  @override
-  CustomeMaterialDesktopPlayOrPauseButtonState createState() =>
-      CustomeMaterialDesktopPlayOrPauseButtonState();
-}
-
-class CustomeMaterialDesktopPlayOrPauseButtonState
-    extends State<CustomeMaterialDesktopPlayOrPauseButton>
-    with SingleTickerProviderStateMixin {
-  late final animation = AnimationController(
-    vsync: this,
-    value: widget.controller.player.state.playing ? 1 : 0,
-    duration: const Duration(milliseconds: 200),
-  );
-
-  StreamSubscription<bool>? subscription;
-
-  @override
-  void setState(VoidCallback fn) {
-    if (mounted) {
-      super.setState(fn);
-    }
-  }
-
-  @override
-  void didChangeDependencies() {
-    super.didChangeDependencies();
-    subscription ??= widget.controller.player.stream.playing.listen((event) {
-      if (event) {
-        animation.forward();
-      } else {
-        animation.reverse();
-      }
-    });
-  }
-
-  @override
-  void dispose() {
-    animation.dispose();
-    subscription?.cancel();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return IconButton(
-      onPressed: widget.controller.player.playOrPause,
-      iconSize: 25,
-      color: Colors.white,
-      icon: AnimatedIcon(
-        progress: animation,
-        icon: AnimatedIcons.play_pause,
-        size: 25,
-        color: Colors.white,
-      ),
-    );
-  }
-}
-
 // BUTTON: VOLUME
 
 /// MaterialDesktop design volume button & slider.
